fix(geometry,ci): split synthesis.rs under its budget; re-bless the watertightness census for the triangle-count gate
Two self-inflicted CI failures from the triangle-count classification
change:

- Module-size ratchet: the new triangle-count comments pushed
  router/voids/synthesis.rs to 973 lines against its 955-line
  allowlist budget. Extracted `remove_internal_membrane` (the AC20
  glued-cap-to-cap deseaming helper) verbatim into a new sibling
  module, router/voids/synthesis/membrane.rs (206 lines, no allowlist
  entry needed). synthesis.rs is now 791 lines.

- Geometry watertightness census: the larger CI corpus (1170 void
  hosts, vs. what's measurable against this repo's own small
  fixtures) contains two openings this metric switch does reclassify.
  Both are in ara3d/ISSUE_126_model.ifc (opening ids #98282 and
  #111997): previously routed to the >100-vertex AABB fallback, they
  now qualify for exact CSG under the >100-triangle gate. Per the
  re-run, both hosts went from torn (26 and 27 unmatched edges) to
  fully watertight (0), with fewer triangles each (220->186, 51->32) -
  the exact path produced a cleaner cut than the fallback did, not a
  worse one. Every other host is unchanged: 0 regressed, 0 coverage
  loss, 0 added, 0 volume-moved, and every corpus total (torn hosts,
  unmatched/strict edges, genuine defects) moved down or held. Golden
  re-blessed for exactly these 2 rows.

Refs #4119
