+++ rust/geometry/tests/manifests/watertightness_census.tsv
@@ -168,9 +168,9 @@ ara3d/ISSUE_126_model.ifc	83694	SweptSolid	0	20	0	0	0	-	0	59590
 ara3d/ISSUE_126_model.ifc	84442	SweptSolid	0	34	0	0	0	-	0	38318
 ara3d/ISSUE_126_model.ifc	86446	SweptSolid	0	54	0	0	0	-	0	141887
 ara3d/ISSUE_126_model.ifc	86609	SweptSolid	0	28	0	0	0	-	0	862544
-ara3d/ISSUE_126_model.ifc	98282	SweptSolid	26	220	0	0	26	0	26	-
+ara3d/ISSUE_126_model.ifc	98282	SweptSolid	0	186	0	0	0	-	0	1776930
 ara3d/ISSUE_126_model.ifc	100504	SweptSolid	0	50	0	0	0	-	0	474969
-ara3d/ISSUE_126_model.ifc	111997	SweptSolid	27	51	0	0	27	0	27	-
+ara3d/ISSUE_126_model.ifc	111997	SweptSolid	0	32	0	0	0	-	0	290818
 ara3d/ISSUE_126_model.ifc	131771	SweptSolid	0	96	0	0	0	-	0	660647
 ara3d/ISSUE_129_N1540_17_EXE_MOD_448200_02_09_11SMC_IGC_V17.ifc	6891	Clipping	0	594	0	0	0	-	0	9688788
 ara3d/ISSUE_129_N1540_17_EXE_MOD_448200_02_09_11SMC_IGC_V17.ifc	7526	SweptSolid	0	2210	0	0	0	-	0	1028308685
